Check if user is still logged in when delete operation is triggered

+++ src/app/components/view-survey/overview/edit-delete/survey-delete/survey-delete.component.ts
@@ -4,6 +4,9 @@ import {matchesSurveyName} from "../../../../../directives/match-survey-name-val
 import {Survey} from "../../../../../model/survey";
 import {SurveyService} from "../../../../../services/survey/survey.service";
 import {Router} from "@angular/router";
+import {KeycloakService} from "keycloak-angular";
+import {User} from "../../../../../model/user";
+import {UserService} from "../../../../../services/user/user.service";
 
 @Component({
   selector: 'app-survey-delete',
@@ -20,7 +23,9 @@ export class SurveyDeleteComponent implements OnChanges {
   }
 
   constructor(private fb: FormBuilder,
+              private keycloak: KeycloakService,
               private surveyService: SurveyService,
+              private userService: UserService,
               private router: Router) {
   }
 
@@ -32,10 +37,26 @@ export class SurveyDeleteComponent implements OnChanges {
   }
 
   deleteSurvey() {
-    this.surveyService.deleteSurvey(this.survey).subscribe(
-      () => {
-        this.router.navigate(['/profile/surveys']);
+    this.keycloak.isLoggedIn().then(isLoggedIn => {
+      if (isLoggedIn) {
+        this.keycloak.loadUserProfile().then(userProfile => {
+          this.userService.findUserByEMail(userProfile.email).subscribe(
+            (response: User) => {
+              if (response.id === this.survey.userId) {
+                this.surveyService.deleteSurvey(this.survey).subscribe(
+                  () => {
+                    this.router.navigate(['/profile/surveys']);
+                  }
+                );
+              } else {
+                console.log('not authorized');
+              }
+            }
+          )
+        })
+      } else {
+        console.log('not authorized');
       }
-    );
+    })
   }
 }
